Refactor Boom API utility functions into a separate module for improved organization and reusability
